+++ src/Janus.Hosting/IdentityEndpoints.cs
@@ -41,6 +41,7 @@ public static class IdentityEndpoints
         _ = endpoints.MapAuthentication();
         _ = endpoints.MapSignOn();
         _ = endpoints.MapAccount();
+        _ = endpoints.MapAccountAdministration();
         _ = endpoints.MapCredentials();
         _ = endpoints.MapRecovery();
         _ = endpoints.MapPrivacy();